* Renamed FFmpegFrameLock to FFmpegLockCallback   * As suggested, don't allocate a Long, but save the lock number directly as value in the (void *).
